shared/django/requirements.txt
djangorestframework-jwt==1.11.0
PyJWT==1.6.4